+++ src/main/resources/inlayProviders/Parameters/hints.lambdas.parameters.py
@@ -1,3 +1,0 @@
-lambda_exp = lambda char, times: char * times
-
-lambda_exp('a', 3)